refactor: 모듈 경로 수정 및 일관성 있는 파일명 변경
